refactor: Relocate ApiResponse and update Repository implementation
This commit introduces the following changes:

- **Data Model:**
    - Moved `ApiResponse.kt` data class into `App.kt`.
- **Networking:**
    - Renamed `ApiService` interface to `RockuApi`.
    - Moved Retrofit instance creation into an `ApiService` object, which now exposes the `RockuApi` instance.
    - Updated `AppUIScreen.kt` and `MainActivity.kt` to use the new `ApiService.BASE_URL` path.
- **Repository:**
    - Deleted the old `Repository.kt` file located in `com.android.rokuapp.data.network`.
    - Created a new `Repository.kt` class in `com.android.rokuapp.data.repository`.
    - The new `Repository` class is now instantiated directly and its `getApps()` method calls the `ApiService.api.getApps()` to fetch data.
- **ViewModel:**
    - In `AppViewModel.kt`, updated the `Repository` import to the new path.
    - Modified `loadApps()` to instantiate `Repository` directly: `Repository().getApps()`.
