+++ app/src/main/java/com/android/rokuapp/data/network/ApiService.kt
@@ -1,21 +1,28 @@
 package com.android.rokuapp.data.network
 
-import com.android.rokuapp.data.model.App
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
+import com.android.rokuapp.data.model.ApiResponse
+import kotlinx.serialization.json.Json
+import okhttp3.MediaType.Companion.toMediaType
+import retrofit2.Retrofit
+import retrofit2.converter.kotlinx.serialization.asConverterFactory
 import retrofit2.http.GET
 
-interface ApiService {
+interface RockuApi {
     @GET("apps.json")
     suspend fun getApps(): ApiResponse
-
-    companion object {
-        const val BASE_URL = "https://rokumobileinterview.s3.us-west-2.amazonaws.com/"
-    }
 }
 
-@Serializable
-data class ApiResponse(
-    @SerialName("apps")
-    val apps: List<App> = emptyList()
-)
+object ApiService {
+    const val BASE_URL = "https://rokumobileinterview.s3.us-west-2.amazonaws.com/"
+
+    private val json = Json { ignoreUnknownKeys = true }
+
+   val api: RockuApi by lazy {
+        Retrofit.Builder()
+            .baseUrl(ApiService.BASE_URL)
+            .addConverterFactory(json.asConverterFactory("application/json".toMediaType()))
+            .build()
+            .create(RockuApi::class.java)
+    }
+
+}
