--- app/src/main/java/com/android/rokuapp/data/network/Repository.kt
@@ -1,23 +1,0 @@
-package com.android.rokuapp.data.network
-
-import com.android.rokuapp.data.model.App
-import com.android.rokuapp.data.network.ApiService.Companion.BASE_URL
-import kotlinx.serialization.json.Json
-import okhttp3.MediaType.Companion.toMediaType
-import retrofit2.Retrofit
-import retrofit2.converter.kotlinx.serialization.asConverterFactory
-
-object Repository {
-
-    private val json = Json { ignoreUnknownKeys = true }
-
-    private val apiService: ApiService by lazy {
-        Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .addConverterFactory(json.asConverterFactory("application/json".toMediaType()))
-            .build()
-            .create(ApiService::class.java)
-    }
-
-    suspend fun getApps(): List<App> = apiService.getApps().apps
-}
